updates to the tests
